Adicionando sweetalert no projeto e ja utilizando no excluir categoria

+++ admin/view/categorias/delete.php
@@ -2,13 +2,22 @@
 
 require_once '../../../autoload.php';
 
-if(!empty($_POST['id'])) {
+if (!empty($_POST['id'])) {
+   
    $id = addslashes($_POST['id']);
    $obj = new Categoria();
-   if($obj->delete($id)) {
-      echo "Categoria excluida com Sucesso";
+   // Verifica se existe um registro com esse id
+   if ($obj->find($id)) {
+      if ($obj->delete($id)) {
+         $json = array("icon" => "success", "title" => "Sucesso!", "text" => "A categoria selecionada foi excluida com Sucesso!");
+      } else {
+         $json = array("icon" => "error", "title" => "Erro!", "text" => "Falha ao excluir categoria");
+      }
    } else {
-      echo "Falha ao excluir categoria";
+      $json = array("icon" => "error", "title" => "Error!", "text" => "Nenhum registro encontrado!");
    }
-   
 }
+
+$jsonString = json_encode($json);
+
+echo $jsonString;
